+++ src/lib/site.ts
@@ -1,7 +1,7 @@
 // Central place for the handful of site-wide constants that show up in the UI
 // and legal copy. Change the contact address here if you'd rather route it
 // somewhere other than the maintainer's personal inbox.
-export const CONTACT_EMAIL = 'sbmagar.sbm99@gmail.com'
+export const CONTACT_EMAIL = 'sagar@sagarbudhathoki.com'
 export const REPO_URL = 'https://github.com/sbmagar13/recipe-jar'
 export const LEGAL_UPDATED = 'July 2026'